// todo-api-client/models/List.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface List
 */
export interface List {
  /**
   *
   * @type {number}
   * @memberof List
   */
  readonly id?: number;
  /**
   *
   * @type {string}
   * @memberof List
   */
  title: string;
  /**
   *
   * @type {boolean}
   * @memberof List
   */
  archived?: boolean;
  /**
   *
   * @type {number}
   * @memberof List
   */
  readonly user?: number;
}

/**
 * Check if a given object implements the List interface.
 */
export function instanceOfList(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "id" in value;
  isInstance = isInstance && "title" in value;
  isInstance = isInstance && "user" in value;

  return isInstance;
}

export function ListFromJSON(json: any): List {
  return ListFromJSONTyped(json, false);
}

export function ListFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): List {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    id: json["id"],
    title: json["title"],
    archived: !exists(json, "archived") ? undefined : json["archived"],
    user: json["user"],
  };
}

export function ListToJSON(value?: List | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    title: value.title,
    archived: value.archived,
  };
}